Added error handling and cascade delete option in employee and device

+++ src/Services/firebase/device-management-firebase/device-management.service.ts
@@ -26,32 +26,54 @@ export class DeviceManagementService {
           this.devices.push(device);
           this.unavailableSerialNumbers.push(devices[key].serialNumber);
         }
-      } 
+      },
+      (error) => {
+        alert("Something went wrong while loading the devices.\nPlease try again later.")
+      }
     );
   }
 
   getDevice(serialNumber:string):void{
     let firebaseKey:string = this.getFirebaseKey(serialNumber);
-    this.http.get("https://capstonedb-a452b-default-rtdb.firebaseio.com/Devices/" + firebaseKey + ".json").subscribe(
+    if(firebaseKey){
+      this.http.get("https://capstonedb-a452b-default-rtdb.firebaseio.com/Devices/" + firebaseKey + ".json").subscribe(
       (device:any)=>{
         let newDeviceSelection = new FirebaseDevice(firebaseKey,device.serialNumber,device.description,device.type);
         this.selectedDevice.emit(newDeviceSelection);
+        },
+        (error)=>{
+          alert("Something went wrong while loading the device.\nPlease try again later.")
         }
-    );
+      );
+    }
+    else{
+      alert("Invalid Device Serial Number.")
+    }
+    
   }
 
   addDevice(device:Device):void{
-    this.http.post("https://capstonedb-a452b-default-rtdb.firebaseio.com/Devices.json", device).subscribe(
+    let firebaseKey:string = this.getFirebaseKey(device.serialNumber);
+    if(!firebaseKey){
+      this.http.post("https://capstonedb-a452b-default-rtdb.firebaseio.com/Devices.json", device).subscribe(
       (firebaseKey:any)=>{
           this.devices.push(new FirebaseDevice(firebaseKey.name,device.serialNumber,device.description,device.type));
           this.unavailableSerialNumbers.push(device.serialNumber);
+        },
+        (error)=>{
+          alert("Something went wrong while adding the device.\nPlease try again later.")
         }
-    );
+      );
+    }
+    else{
+      alert("The device Serial Number is already in the database.\nYou cannot add it again.")
+    }
   }
 
   editDevice(deviceToEdit:Device):void{
     let firebaseKey:string = this.getFirebaseKey(deviceToEdit.serialNumber);
-    this.http.put("https://capstonedb-a452b-default-rtdb.firebaseio.com/Devices/" + firebaseKey + ".json", deviceToEdit).subscribe(
+    if(firebaseKey){
+      this.http.put("https://capstonedb-a452b-default-rtdb.firebaseio.com/Devices/" + firebaseKey + ".json", deviceToEdit).subscribe(
       (device:any)=>{
         this.devices.splice(this.devices.findIndex(
           (device:FirebaseDevice)=>{
@@ -59,13 +81,21 @@ export class DeviceManagementService {
           }),1);
 
         this.devices.push(new FirebaseDevice(firebaseKey,device.serialNumber,device.description,device.type));
+        },
+        (error)=>{
+          alert("Something went wrong while editing the device.\nPlease try again later.")
         }
-    );
+      );
+    }
+    else{
+      alert("The device you are trying to edit does not exist.\nPlease check the Serial Number.")
+    }
   }
 
   deleteDevice(serialNumber:string):void{
     let firebaseKey:string = this.getFirebaseKey(serialNumber);
-    this.http.delete("https://capstonedb-a452b-default-rtdb.firebaseio.com/Devices/" + firebaseKey + ".json").subscribe(
+    if(firebaseKey){
+      this.http.delete("https://capstonedb-a452b-default-rtdb.firebaseio.com/Devices/" + firebaseKey + ".json").subscribe(
       ()=>{
         this.devices.splice(this.devices.findIndex(
           (device:FirebaseDevice)=>{
@@ -76,17 +106,30 @@ export class DeviceManagementService {
               return serialNumberUnavailable === serialNumber;
             }
           ),1);
-      }
-    );
+        },
+        (error)=>{
+          alert("Something went wrong while deleting the device.\nPlease try again later.")
+        }
+      );
+    }
+    else{
+      alert("The device you are trying to delete does not exist.\nPlease check the Serial Number.")
+    }
   }
 
   private getFirebaseKey(serialNumber:string):string {
-    let firebaseKey:string = this.devices.filter(
-      (device:FirebaseDevice)=>{
-        return device.serialNumber === serialNumber
-      })[0].firebaseKey;
-    return firebaseKey;
+    let deviceIndex:number = this.getIndexOfFirebaseKey(serialNumber);
+    if(deviceIndex === -1)
+      return "";
+    return this.devices[deviceIndex].firebaseKey;
   }
-
+  
+  private getIndexOfFirebaseKey(serialNumber:string):number{
+    return this.devices.findIndex(
+      (device:FirebaseDevice) => {
+        return device.serialNumber === serialNumber;
+      }
+    );
+  }
 
 }
